Crate juniper_codegen

This crate supplies custom derive implementations for the juniper crate.

You should not depend on juniper_codegen directly. You only need the juniper crate.

Attribute Macros

object
The object proc macro is the primary way of defining GraphQL resolvers that can not be implemented with the GraphQLObject derive.

Derive Macros

GraphQLEnum

GraphQLInputObject

GraphQLObject

GraphQLScalarValue
This custom derive macro implements the #[derive(GraphQLScalarValue)] derive.

ScalarValue	Deprecated